[FIX] warnings
git-svn-id: https://coolfluidsrv.vki.ac.be/svn/coolfluid3/trunk@1941 4892e656-8fb9-448a-b5bc-b1919152596f

+++ src/Common/MacOSX/OSystemLayer.cpp
@@ -177,7 +177,7 @@ feenableexcept (unsigned int excepts)
   fenv.__control &= ~new_excepts;
   fenv.__mxcsr   &= ~(new_excepts << 7);
 
-  return ( fesetenv (&fenv) ? -1 : old_excepts );
+  return ( fesetenv (&fenv) ? -1 : (int)old_excepts );
 }
 
 #if 0 // not used for the moment
